Upgrade bindgen to 0.31.3 and bump our version to 0.2.0
Changes to generated code yield breaking changes in our library, so we
must bump versions.

We blacklist IRQn_Type and provide our own implementation

IRQn_Type is an enum in components/device/nrf52.h

We need the members to be Rust i8, but this conflicts with bindgen's
requirement to get the enum's type from ctypes::c_schar a u8.

Changing our c_schar to i8 breaks a lot of stuff, so we can provide our
own type alias to an i8, and all the constants.

Other changes:
* gitignore vscode files, Cargo.lock
* Remove unused macro_use and imports
* some intermediate modules are now not generated, so we must access
members and names from the nrf module.